pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["orpheusplus*"]

[project]
name = "orpheusplus"
version = "0.0.1"
authors = [
  { name = "Zheng-Xiang Ye", email="r12b48005@ntu.edu.tw"},
  { name = "Tsai, Chih-Yun", email="r12631014@ntu.edu.tw"},
  { name = "Bai, Ci-Ruei", email="r12631025@ntu.edu.tw"},
  { name = "Chen, Ssu-Chi", email="r12631036@ntu.edu.tw"},
  
]
description = "TODO"
readme = "README.md"
dependencies = [
    "mysql-connector-python==8.0.33",
    "pytest",
    "maskpass",
    "pathlib",
    "PyYAML",
    "networkx[default]",
    "sqlparse",
    "tabulate",
    "wcwidth",
    "keyring"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
orpheusplus = "orpheusplus.cli:main"